### 2016-05-05 #12 @whedon
Submission: mst_clustering: Clustering via Minimum Spanning Trees
**Submitting author:** @jakevdp (<a href="http://orcid.org/0000-0002-9623-3401">Jacob VanderPlas</a>)
**Repository:** <a href="http://github.com/jakevdp/mst_clustering" target="_blank">http://github.com/jakevdp/mst_clustering</a>
**Version:** v1.0.0
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.50995" target ="_blank">http://dx.doi.org/10.5281/zenodo.50995</a>

- [x] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.5281/zenodo.50995">software archive</a> resolve?

- [x] **Performance:** Have the performance claims of the software been confirmed?

- [x] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?
