### 2016-05-05 #12 @arfon
> I notice that the generated PDF doesn't have any link or reference to github or the software DOI – are authors meant to put that in explicitly? I would have expected it to be included automatically somehow.

Good catch. I've been meaning to add that but completely forgot. Thanks for the reminder!
