### 2016-05-07 #12 @arfon
/ cc @openjournals/joss-reviewers anyone interested in reviewing this?
