### 2016-05-07 #12 @nicoguaro
I am writing this _review_ following the points stated above. The software provides the implementation of a particular algorithm for clustering.

# General checks

- License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?

    Yes, the license is [2-clause BSD license](https://en.wikipedia.org/wiki/BSD_licenses#2-clause_license_.28.22Simplified_BSD_License.22_or_.22FreeBSD_License.22.29). I think it can be helpful to provide this information explicitly.

- Version: Does the release version given match the GitHub release (v1.0.0)?

    No. The version that is installed when I run `pip install mst_clustering` is 0.1.

- Archive: Does the software archive resolve?

    Yes.

# Functionality

- Installation: Does installation proceed as outlined in the documentation?

    Yes.

- Have the functional claims of the software been confirmed?

    Yes.

- Performance: Have the performance claims of the software been confirmed?

    I could not find a claim about performance.

# Documentation

- A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

    Partially. It does state what the algorithm does, but it does not say what the target audience is.

- Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

    Yes.

- Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

    Yes. The repository includes an IPython notebook with this purpose.

- Functionality documentation: Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?

    Partially. Besides the examples provided in the IPython notebook there are not more documentation and functions do not have `docstrings`.

- Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

    The tests run ok using `nosetests test_mst_clustering.py`. But they make a relative reference instead of instead of importing the package. I suggest to change line 11 of this [script](https://github.com/jakevdp/mst_clustering/blob/master/mst_clustering/tests/test_mst_clustering.py).

- Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

    No. Recommend editing README.md to make clear. It does include a description about [How to make new releases](https://github.com/jakevdp/mst_clustering/blob/master/HOWTO_RELEASE.md).

# Software Paper

- Authors: Does the paper.md file include a list of authors with their affiliations?

    Yes.

- A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

    Partially. It does state what the algorithm does, but it does not say what the target audience is.

-References: Do all archival references that should have a DOI list one (e.g. papers, datasets, software)?

    I am not sure to understand this point, but none of the references have a DOI.

Since I am not sure about the procedures I do not know if it should be recommended to continue the process.
