### 2016-05-08 #12 @arfon
@nicoguaro - many thanks for this review :zap: I'll let @jakevdp reply to some of your questions here.

> References: Do all archival references that should have a DOI list one (e.g. papers, datasets, software)?

This questions is about items in the references. Published papers and datasets will usually have a DOI and so either are cited then we would like the author to include the DOI in the reference too.

> Since I am not sure about the procedures I do not know if it should be recommended to continue the process.

@nicoguaro - is it safe to assume that if your comments are addressed then you recommend this for publication?

@jakevdp - it looks like there are a couple of quick things to address here.
