### 2016-05-09 #12 @jakevdp
@nicoguaro – thanks for the review! I'll try to address everything you brought up.

One question regarding documentation: you mentioned that functions do not have docstrings, but the class and both its methods defined by the module [*do* have doc-strings](https://github.com/jakevdp/mst_clustering/blob/master/mst_clustering/_mst_clustering.py). Can you point me to what was specifically concerning you?

Also, since this is basically a module with a single method, I felt that an example notebook would be much more useful than a typical sphinx/readthedocs-style website. Do you feel the current setup is sufficient for a potential user, or should I do the extra work to create a separate docs website?

Thanks
