### 2016-05-09 #12 @nicoguaro
@jakevdp – thanks for taking the comments into consideration.

Regarding the documentation, I apologize, I see the docstrings now. I use Spyder regularly, and the `Ctrl + i` hotkey to render docstrings through `sphinx`... but it did not work before. But the docstrings are there and they can be pulled.

I also agree that the example notebook is good enough, maybe you would like to compare the results with some other method, but that is just a suggestion.
