### 2016-05-12 #12 @arfon
@jakevdp @nicoguaro - just checking in on this.
