### 2016-05-13 #12 @jakevdp
I need to address comments. I've been wrapped up with finishing and astropy PR by tomorrow to get it into v1.2 :smile:
