### 2016-05-13 18d51e4
DOC: fix image in README
