### 2016-05-13 b8f71cf
DOC: clarifications in README
