### 2016-05-13 c66f353
add community guidelines to README
