### 2016-05-13 4df79c5
TST: use absolute rather than relative import
