### 2016-05-13 d987afa
MAINT: include README in manifest
