### 2016-05-13 9523e40
MAINT: bump version to 1.0
