### 2016-05-13 342dddf
MAINT: bump release to 1.1.dev0
