### 2016-05-13 #12 @arfon
> OK – I've pushed updates to the mst_clustering repository addressing all of @nicoguaro's concerns. What's the next step @arfon?

@nicoguaro - are we good to accept?
