### 2016-05-16 #12 @arfon
> @arfon - We are good to accept. I just installed it in a different computer to check, and everything seems alright.

👍
