### 2016-05-16 #12 @arfon
Thanks all! http://dx.doi.org/10.21105/joss.00012
